+++ Asteroids Reimagined/Enemy.h
@@ -30,8 +30,8 @@ public:
 	bool Initialize(Utilities* utilities);
 	bool BeginRun();
 
-	void Update(double deltaTime);
-	void FixedUpdate(double deltaTime);
+	void Update(float deltaTime);
+	void FixedUpdate(float deltaTime);
 	void Draw3D();
 
 	bool CheckRockCollisions(Entity* rock);
